requirements.txt
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1743597017592/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1734342155601/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1742243108/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1691763562544/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560578465/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
dataclasses-json @ file:///home/conda/feedstock_root/build_artifacts/dataclasses-json_1734217732310/work
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
filelock==3.18.0
frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1737645287563/work
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet @ file:///Users/runner/miniforge3/conda-bld/greenlet_1734532814058/work
grpcio==1.71.0
grpcio-status==1.71.0
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1733327467879/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1731707562/work
httplib2==0.22.0
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
httpx-sse @ file:///home/conda/feedstock_root/build_artifacts/httpx-sse_1734732046715/work
huggingface-hub==0.30.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
Jinja2==3.1.6
jiter @ file:///Users/runner/miniforge3/conda-bld/jiter_1741655126008/work
joblib==1.4.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302931775/work
langchain @ file:///home/conda/feedstock_root/build_artifacts/langchain_1743828705034/work
langchain-community @ file:///home/conda/feedstock_root/build_artifacts/langchain-community_1743830692417/work
langchain-core @ file:///home/conda/feedstock_root/build_artifacts/langchain-core_1743794189038/work
langchain-text-splitters @ file:///home/conda/feedstock_root/build_artifacts/langchain-text-splitters_1743799028984/work
langgraph @ file:///home/conda/feedstock_root/build_artifacts/langgraph_1743724686578/work
langgraph-checkpoint @ file:///home/conda/feedstock_root/build_artifacts/langgraph-checkpoint_1742285937817/work
langgraph-prebuilt @ file:///home/conda/feedstock_root/build_artifacts/langgraph-prebuilt_1743747557261/work
langgraph-sdk @ file:///home/conda/feedstock_root/build_artifacts/langgraph-sdk_1743664233648/work
langsmith @ file:///home/conda/feedstock_root/build_artifacts/langsmith_1737099297406/work
MarkupSafe==3.0.2
marshmallow @ file:///home/conda/feedstock_root/build_artifacts/marshmallow_1738612362469/work
mpmath==1.3.0
msgpack @ file:///Users/runner/miniforge3/conda-bld/msgpack-python_1725975043881/work
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1743843481548/work
mypy_extensions @ file:///home/conda/feedstock_root/build_artifacts/mypy_extensions_1733230897951/work
networkx==3.4.2
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225421156/work/dist/numpy-1.26.4-cp312-cp312-macosx_11_0_arm64.whl#sha256=011d57633d659db8280a5811d62d0c5a615719a6d7cbd46a04f6ffce0c2a7db3
openai @ file:///home/conda/feedstock_root/build_artifacts/openai_1743467416804/work
orjson @ file:///Users/runner/miniforge3/conda-bld/orjson_1742909947333/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pillow==11.1.0
pinecone==6.0.2
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
propcache @ file:///Users/runner/miniforge3/conda-bld/propcache_1737635647354/work
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1743701108924/work
pydantic-settings @ file:///home/conda/feedstock_root/build_artifacts/pydantic-settings_1740671914185/work
pydantic_core @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_pydantic-core_1743607651/work
pyparsing==3.2.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1742948348/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1737454693669/work
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
requests-toolbelt @ file:///home/conda/feedstock_root/build_artifacts/requests-toolbelt_1733734787568/work
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==78.1.0
six==1.17.0
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
SQLAlchemy @ file:///Users/runner/miniforge3/conda-bld/sqlalchemy_1743109654644/work
sympy==1.13.1
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.6.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers==4.35.2
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1741438046699/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1743820059/work
typing_inspect @ file:///home/conda/feedstock_root/build_artifacts/typing_inspect_1733845867367/work
uritemplate==4.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
wheel==0.45.1
xxhash @ file:///Users/runner/miniforge3/conda-bld/python-xxhash_1740594854084/work
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1737575870898/work
zstandard==0.23.0
